Refactored the `check` code to make it more compact and less usage of duplicate code
